Clean up AntiSpam code
