+++ src/main/java/net/wurstclient/hacks/AntiSpamHack.java
@@ -12,15 +12,14 @@ import java.util.List;
 import net.minecraft.client.gui.hud.ChatHud;
 import net.minecraft.client.gui.hud.ChatHudLine;
 import net.minecraft.client.util.ChatMessages;
-import net.minecraft.text.CharacterVisitor;
 import net.minecraft.text.MutableText;
 import net.minecraft.text.OrderedText;
-import net.minecraft.text.Style;
 import net.minecraft.util.math.MathHelper;
 import net.wurstclient.Category;
 import net.wurstclient.SearchTags;
 import net.wurstclient.events.ChatInputListener;
 import net.wurstclient.hack.Hack;
+import net.wurstclient.util.ChatUtils;
 import net.wurstclient.util.MathUtils;
 
 @SearchTags({"NoSpam", "ChatFilter", "anti spam", "no spam", "chat filter"})
@@ -51,33 +50,6 @@ public final class AntiSpamHack extends Hack implements ChatInputListener
 		if(chatLines.isEmpty())
 			return;
 		
-		/**
-		 * A {@link CharacterVisitor} to completely bypass Mojang's visitor
-		 * system and just get the damn {@link String} out of a
-		 * {@link ChatHudLine}.
-		 *
-		 * <p>
-		 * Is this seriously the replacement for <code>getString()</code>?
-		 * What were they thinking?!
-		 */
-		class JustGiveMeTheStringVisitor implements CharacterVisitor
-		{
-			StringBuilder sb = new StringBuilder();
-			
-			@Override
-			public boolean accept(int index, Style style, int codePoint)
-			{
-				sb.appendCodePoint(codePoint);
-				return true;
-			}
-			
-			@Override
-			public String toString()
-			{
-				return sb.toString();
-			}
-		}
-		
 		ChatHud chat = MC.inGameHud.getChatHud();
 		int maxTextLength =
 			MathHelper.floor(chat.getWidth() / chat.getChatScale());
@@ -89,17 +61,12 @@ public final class AntiSpamHack extends Hack implements ChatInputListener
 		
 		for(int i = chatLines.size() - 1; i >= 0; i--)
 		{
-			JustGiveMeTheStringVisitor oldLineVS =
-				new JustGiveMeTheStringVisitor();
-			chatLines.get(i).content().accept(oldLineVS);
-			String oldLine = oldLineVS.toString();
+			String oldLine = ChatUtils.getAsString(chatLines.get(i));
 			
 			if(matchingLines <= newLines.size() - 1)
 			{
-				JustGiveMeTheStringVisitor newLineVS =
-					new JustGiveMeTheStringVisitor();
-				newLines.get(matchingLines).accept(newLineVS);
-				String newLine = newLineVS.toString();
+				String newLine =
+					ChatUtils.getAsString(newLines.get(matchingLines));
 				
 				if(matchingLines < newLines.size() - 1)
 				{
@@ -119,10 +86,8 @@ public final class AntiSpamHack extends Hack implements ChatInputListener
 				
 				if(i > 0 && matchingLines == newLines.size() - 1)
 				{
-					JustGiveMeTheStringVisitor nextOldLineVS =
-						new JustGiveMeTheStringVisitor();
-					chatLines.get(i - 1).content().accept(nextOldLineVS);
-					String nextOldLine = nextOldLineVS.toString();
+					String nextOldLine =
+						ChatUtils.getAsString(chatLines.get(i - 1));
 					
 					String twoLines = oldLine + nextOldLine;
 					String addedText = twoLines.substring(newLine.length());
